Really don't break PGP, and don't break HTML messages.
For PGP, we need to upgrade the entire message from multipart/signed
to multipart/mixed, and the multipart/signed part and its subparts
will be the first mixed part, with our added text/html part as
the second.

For HTML, if there were unclosed HTML tags, HTML::TreeBuilder would
destroy the message. Set implicit_tags to false so that it (mostly)
leaves broken trees alone but still allows us to rewrite them.