Refresh subparts depth-first so the final message gets all changes
Email::MIME body_str_set doesn't necessarily update the cache so we must
force it to refresh.

We do this by forcing it to recalculate any parts that have subparts, but
we must do this depth first, so wrapping multipart parts will get updated
information.

For example, if our message looks like:

          A
        / | \
       B  C  D
            / \
           E   F
              / \
             G   H

We must rebuild F so that when we rebuild D it gets all changes. We must
rebuild D before rebuilding A so that it gets all changes. Etc...